eeze: Fix eeze_udev_find_by_subsystem_sysname function
This function was severly broken when used for finding devices of a
particular subsystem. Previously, this function would return every
device in the system, regardless if subsystem or sysname was passed in.
This patch fixes the issue by adding udev_enumerate matches before we
scan for devices. This way, when we enumerate the scan results, we now
get only devices that match the subsystem (if supplied).

@fix

Signed-off-by: Chris Michael <cpmichael@osg.samsung.com>