Elm_Entry: Magnifier Not working properly on corner areas issue fix.
Summary:
When magnifier is moved across corners and outside entry, working of magnifier
is not proper.

Magnifier is Restricted with in the window and to entry area and its position depends
on cursor.
Signed-off-by: Thiep Ha <thiepha@gmail.com>
Signed-off-by: Shilpa Singh <shilpa.singh@samsung.com>

@fix

Test Plan:
1. Enable Magnifier export ELM_MAGNIFIER_ENABLE=1
2. Launch elementary test and launch entry3 demo
3. Move magnifier on corners and outside of second entry

The contents of magnifiers are undefined when moved outside entry
and clipper moves in a different direction as magnifier.

Reviewers: cedric, raster, thiepha

Reviewed By: thiepha

Subscribers: buds, govi, rajeshps, subodh6129

Differential Revision: https://phab.enlightenment.org/D3586