Scaling test: reorder instructions to set the correct scale
If the scale is set on an object before contents are set, it will not
pass to them. Because of this, in the test, scale of the first label
remains 1.0, i.e the window scale, instead of 0.5.
The patch modifies the order of the instructions by setting the scale
after setting the label as content of the frame.