Revert "eina: rely on GNU make to find file at their correct location."
This reverts commit e26fcbb1dc588c5130e477bf832c59386b1c1951.

This reverts one of three parts of the try to get sub directory
compilation back into eina. It breaks our distcheck though and I
talked to Cedric about it and he prefers to revert these as we might
need to go another route to bring this functionality back. Details
will come to the mailing list.