Revert "eina: reintroduce additional support for eina sub directory compilation."
This reverts commit 1affc60d00e1fa2b702d6b170514b86e1438ed9a.

This reverts one of three parts of the try to get sub directory
compilation back into eina. It breaks our distcheck though and I
talked to Cedric about it and he prefers to revert these as we might
need to go another route to bring this functionality back. Details
will come to the mailing list.