tests eeze: enable eeze_net IPv6 test again and skip if impossible
These test depends on the host system having IPv6 setup on the loopback device.
This is not always the case even if we have a system that does have
"struct ipv6_mreq", our HAVE_IPV6 test and would support it.

Skip here if we can't fetch an IPv6 address. The IPv4 test is still running
regardless.