Add promise
Add a promise object that allows Eolian interface to include promises
as a way to have asynchronous value return and composibility.

The usage is like this in a .eo file:

class Foo {
   methods {
      bar {
         params {
            @inout promise: Promise<int>;
         }
      }
  }
}

Which will create the following API interface:

void foo_bar(Eo* obj, Ecore_Promise** promise);

and the equivalent declaration for implementation.

However, the API function will instantiate the Promise for the user
and the implementer of the class automatically. So the user of this
function will treat it as a @out parameter, while the developer of the
function will treat it like a @inout parameter.

So, the user will use this function like this:

Ecore_Promise* promise; // No need to instantiate
foo_bar(obj, &promise);

int* p = ecore_promise_value_get();
or:
ecore_promise_then(promise, callback);