docs: bring back docs from eo files by switching to the new group name
During the migration of docs to the eo file the group name changed to have a
Elm_ prefix which resulted in our docs not including this group and not showing
any of the docs in there. Switching to the new name fixes this.

Having this prefix makes sense especially as we are now merging elm into efl.

@fix

Fixes T3109