Evas GL shaders: Add random noise in the fragment shader
This small noise, added to the fragment shader's output color, makes
smooth up-scaling of gradient images look very good. Gradient images
that are zoomed in a lot show some banding or blocks, that are
the dithering pixels (good for the original size, terrible when
upscaled).

The simplest solution is to add noise, not even using a proper
dithering pattern. This patch adds noise everytime there is a
texture, without any smart logic. Still looks beautiful.