edje_cc: fix misleading error
edje_cc tries several images paths when searching for an image.
It stops its search only when the image was successfully loaded
or all paths have been tested.

If a load failed for some reason, but the file DOES exist, the
relevant error, that contains the real reason of the failure will
be overwriten by what is very likely to be "File Not Found" since
another path has been tested, due to the failure of the previous
one.

In such case, the user received the "File Not Found" error, which
is not the reason of the failure.

@fix