Revert "edje: seems like preloading is fixed and we can now run edje_cc in //"
This reverts commit e6a27e13b54a024ca849b9fd798398d51f2e27f8.

While preloading itself is fixed, we forgot a little detail there is limit to
the amount of files we can open on any system. This is high enough on Linux that
we never see the issue, but on MacOS X, we can't even build elementary theme.

Reverting this for now and until we have a proper fix for this.