docs: make elementary docs available form main doc page
While elm docs have been merged in they never have been accessible from the
main page of our docs. Fix this by including elm_intro.h and referencing it
from the main page. While we are at it rename it to the scheme we used in EFL.

Another missing piece was the index with all widget references. Never brought
over form the elm legacy tree. Putting this in and we are now at least able to
navigate to all elm widget pages.