Fix Makevars and update all pos'
In the elementary tree all translatable messages are enclosed in a E_() macro, and since this macro was undefined in Makevars all massages was not included in catalog.
If some coder think that the E domain is useless, or deprecated or whatever, please remove it and update all source files accordingly.