evas - legacy evas_object_del - always hide obj regardless of refs
if an object iot reffed or not hide on del. it should have been this
way before eo. eoifications i think messed a few things up.

this does bring up an issue... in eo we have no way to explicitly do
stuff on eo_del regardless of references at the time. this needs to be
solved.

@fix