edje: fix the source string for table item
If the table item has a name, posible case when item name length, with
index, will be 12. The 12 is predefined length for box index.

Quote Cedric
In a box, the index is one dimension, one int, thus the length
of it (from int to string) will always fit inside 12 bytes. That's
where this 12 comes from. That's also how the unique name of that item
is defined.

This commit separate the items name generation by part type, it will be
more correctly.

@fix